Hold TetGen's pointers in a type wide enough for them
TetGen aligns the sentinel records it compares every adjacency against -
dummytet and dummysh - and its point sampling blocks by casting the
pointer to unsigned long, rounding it up, and casting back. unsigned long
is 64 bits where TetGen has been used for twenty years and 32 bits on
Win64, so on Windows those pointers came back truncated to their low
half and every comparison against them, and every walk from them, ran off
into memory that was never theirs.

The effect is not a wrong mesh but an access violation inside
tetrahedralize, on any input at all: a four-point piecewise linear
complex describing a single tetrahedron is enough. Nothing that reaches
TetGen has ever worked in a Windows build - vmtktetgen and
vmtkmeshgenerator included.

The casts now go through uintptr_t, which tetgen.h already typedefs for
this compiler. The remaining unsigned long casts of pointers are all
inside printf diagnostics, where the worst they do is print half an
address.

+++ vtkVmtk/Utilities/tetgen1.4.3/tetgen.h
@@ -3300,7 +3300,7 @@ inline bool tetgenmesh::isfacehasedge(face* s, point tend1, point tend2) {
 
 inline bool tetgenmesh::issymexist(triface* t) {
   tetrahedron *ptr = (tetrahedron *) 
-    ((unsigned long)(t->tet[t->loc]) & ~(unsigned long)7l);
+    ((uintptr_t)(t->tet[t->loc]) & ~(uintptr_t)7l);
   return ptr != dummytet;
 }
